Split regime stress supportability
